JINRO_PROJ/ai_server/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.30.1
python-multipart==0.0.9

pydantic==2.7.1
email-validator==2.2.0
python-dotenv==1.0.1

opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4
Pillow==11.1.0
mediapipe==0.10.14

torch
torchvision
faster-whisper==1.1.1
openai==1.68.2
httpx==0.28.1
requests==2.32.3
